Pass new window and new pane commands directly as parameters to new-window and splitw.
The advantage of this is that exiting or stopping the window/pane command either kills the window/pane or places it in "dead" mode, depending on the remain-on-exit [off | on] session setting. Dead windows/panes can be then killed or respawned with a single command or key-binding. This effectively enables the user to create command-specific views that can be easily refreshed with one keystroke and remain neat by making it impossible to exit to the shell prompt.