Allow base index to be configurable.
Without this fix, only a base index of 0 works.

It looks like 8a91b708cb00ac7364f422c5bb94fc14db62b9b9 imposed the
window counting at 0, but it feels like this should be flexible.

The simple solution I put in here requires the user to specify the base
index, but it might also be possible to read it from tmux also. I
couldn't find a good example of that, however, so I left it as is.