Add local project file (`pwd`/.tmuxinator.yml) support.
We have a need to be able to commit project files to git and share them
between machines (we use disposbale development boxes).  Refactored
`Tmuxinator::Config.project` to also check for `$PWD/.tmuxinator.yml`
when searching for project files.  This occurs between the recursive
search of `Tmuxinator::Config.root` and the default path being selected,
so as to allow for project files to be shared with no disruption to
existing functionality.