Adds a default argument for a --default [filename].yml to be used as the template.
The use case for this is when having muliple simular projects in your
daily tasks that can't inherit from only one template type.

The approach used was to dry up code by removing default? which is
essentially exists?(filename) but without accepting the argument.
Therefore this method could be removed. default(filename) was adjusted
to accept a filename parameter which still defaults to "default" if one
isn't passed.  Finally if filename and default do not exist then the
sample method should be returned.  This also removed the need for
calling to_sym as the default already returns the appropriate file
path.