Add tmuxinator option to specify startup_pane.
Fixes #380

Before this change it was only possible to select the startup_window.
This change introduces an additional option 'startup_pane'. It can be
used to select the pane of the selected window on startup. The index
uses 0...x. Consequenty only panes of the selected window
('startup_window') can be specified.