Preserve existing Synchronize behaviour
This change alters preserves the existing synchronize behaviour. By passing in
'before' the command to synchronize panes will be given before running the pane
commands. This is also the default for existing projects which have `synchronize: true`
set.

By passing 'after', the revised behaviour of giving the synchronize command after
the pane commands will be used.