add hooks to project; close #476
As suggested in #476, would be useful to be able to specify command at
different stages of tmux session. Hooks are an effective way to execute
specific commands when specific events occurs. In particular it adds
on_detach, on_reattach, and on_stop hooks, to terminate or bootstrap
processes and daemons on session events.