Run stop hook before killing the session
Changes the template for the exec ran when we run tmuxinator stop to run
the stop hook before killing the tmux session.

This way if we use tmuxinator stop inside the tmux session itself, the
hook will run fine and the session will close shortly thereafter.

If we run tmuxinator stop from outside the session, the stop hook will
run and then the session will be killed, essentially giving us the same
result as before but allowing the session to be closed from inside
itself (as long as we do it with the tmuxinator stop command).