Extract version check from project
The version check was bundled with the dependency checks. Only the
latter depend on the project, so the version checking was extracted into
a new module.

This module is intended for all things related to the tmux version. So
it made sense to move the supported versions constant there as well.