Add quiet flag to prevent tmux version warning (#583) (#659)
* Extract version check from project

The version check was bundled with the dependency checks. Only the
latter depend on the project, so the version checking was extracted into
a new module.

This module is intended for all things related to the tmux version. So
it made sense to move the supported versions constant there as well.

* Renamings and cleanup for tmux version module

* Renamed tmux version warning message
  This message is visible for the master version AND for very old tmux
  versions. Therefore, it is neither a master nor a deprecation message.
  It is simply a message for unsupported versions.

* Removed negation
  "Supported if included" read easier than "Unsupported if not included"

* Shortened names in the context of the new module

* Add test for version warning message

* Change version warning message to WARNING

* Add test for supported version check

* Add unsupported version warning for stop cmd

* Add unsupported version warning for local cmd

* Add flag to prevent tmux version warning (#583)

This adds the "--suppress-tmux-version-warning" flag to skip the warning
for unsupported tmux versions.

We decided that "--quiet" as a name would be too generic, because that
would imply _all_ messages get hidden, including the deprecation
warnings.

* Remove duplication for the continuation prompt