Fix failing tests when TMUXINATOR_CONFIG is set
When the environment variable TMUXINATOR_CONFIG is set to a custom path,
the specs would fail*. This fixes the affected tests.

This is more of a bandaid to get the tests working because it does not
change the fact that there is stubbing at many different levels:

- Entire methods
- Parts of methods, especially File.directory?
- Env variables

Stubbing ENV is not ideal, see for example
https://robots.thoughtbot.com/testing-and-environment-variables
But in this case, I get the feeling that setting up a cleanly stubbed
ENV would replace most of the other stubbing while being easier to
understand. This may be an idea for the future.

*Reproduce with:
TMUXINATOR_CONFIG=$(pwd)/spec/fixtures/dot-tmuxinator bundle exec rspec spec