Improve consistency in tests
The class being tested was sometimes referred to by it's name and
sometimes the described_class feature of rspec was used. This changes
all occurrences to use the latter.

I personally think this version has some advantages, like easier
renaming of classes and improved readability (the keyword gets
highlighted and different specs look the same).

But more importantly, picking one version and sticking with it makes the
specs more consistent.