Fix failing spec with custom TMUXINATOR_CONFIG path (#666)
* Fix failing tests when TMUXINATOR_CONFIG is set

When the environment variable TMUXINATOR_CONFIG is set to a custom path,
the specs would fail*. This fixes the affected tests.

This is more of a bandaid to get the tests working because it does not
change the fact that there is stubbing at many different levels:

- Entire methods
- Parts of methods, especially File.directory?
- Env variables

Stubbing ENV is not ideal, see for example
https://robots.thoughtbot.com/testing-and-environment-variables
But in this case, I get the feeling that setting up a cleanly stubbed
ENV would replace most of the other stubbing while being easier to
understand. This may be an idea for the future.

*Reproduce with:
TMUXINATOR_CONFIG=$(pwd)/spec/fixtures/dot-tmuxinator bundle exec rspec spec

* Improve consistency in tests

The class being tested was sometimes referred to by it's name and
sometimes the described_class feature of rspec was used. This changes
all occurrences to use the latter.

I personally think this version has some advantages, like easier
renaming of classes and improved readability (the keyword gets
highlighted and different specs look the same).

But more importantly, picking one version and sticking with it makes the
specs more consistent.