Feature/Request Development environment
While developing for tmuxinator i got increasingly frustrated with the
fact that i couldn't properly recreate bugs or environments.

So this is my attempt to fix it.

- Adds vagrant support for multi distro testing
- Adds docker support

This creates a way to develop localy more easily. Please read the
development section in CONTRIBUTING.md on how this all functions.

I also added to the gitignore file for some related and commonly used
files by developers.