Add config-type argument to start command, fix bug, fix specs
You can now pass in a config type to the start command. "default" and
"procfile" are now alpha supported.

Fixed a bug for a mis-named variable.

Comment a spec as it is now a private method in a function. It will be
reimplemented in behavioral tests when the class tests are written.