Move Project.load out of the Config.validate method
This separates the concerns of locating and parsing the config file.

This will allow us to better isolate the Config class as a whole.

Config.validate now returns a string to the location of the config file.

Consumers of Config.validate are now responsible for executing their own
Project.load.