docs: fish completions file not downloaded to correct directory
The command given in the README.md downloads the fish file to the current working directory instead of to the provided directory, because the `-O` option is left out. Additionally, the `-O` option's value must a file, not a directory.

Having the `$` sign at the start of the given command is inconvenient, because, after copying and pasting, one has to firstly delete it.