fix: Properly pass args to start when no command is given
Problem

mux start [project] [args]

should behave the same as:

mux [project] [args]

Currently, the args are not passed through to the start command as
Thor `options` - they come through as args instead.

Solution

Fix the call to Tmuxinator::Cli#start when a project name is passed in
by simply appending `:start` to the args, as the first argument is the
subcommand in Thor::Cli.

Signed-off-by: Andrew Kofink <ajkofink@gmail.com>